Make File
This program was created to create a new file within finder. Ever wanted to create a file within a directory similar to windows yet, OSX doesn't have that ability, well this program takes care of that issue. Simple add this program to your finder program and you'll be able to make files within the current directory your browsing. Please feel free to use.